Fix pipeline
