+++ src/dev/src/app/image-gallery/image-gallery-in-modal.demo.html
@@ -5,9 +5,12 @@
 -->
 
 <h2>Image Gallery in Modal Demo</h2>
+
+<!-- ── Fixed size ─────────────────────────────────────────────────────────── -->
+<h3>Fixed image size (<code>[imageSize]="400"</code>)</h3>
 <p>
   The gallery is embedded inside a Clarity modal with <code>[useBuiltInModal]="false"</code>. Clicking an image replaces
-  the modal content with the carousel instead of opening a nested modal.
+  the modal content with the carousel instead of opening a nested modal. The carousel uses explicit pixel sizes.
 </p>
 
 <button class="btn btn-primary" (click)="openModal()">Open Modal</button>
@@ -58,6 +61,63 @@
       [thumbnailSize]="48"
       [hideHeader]="true"
       (closeCarousel)="onCarouselClose()"
+      style="flex: 1; min-height: 0; display: flex; flex-direction: column"
+    ></clr-image-carousel>
+  </div>
+</clr-modal>
+
+<!-- ── Auto size ──────────────────────────────────────────────────────────── -->
+<h3 style="margin-top: 2rem">Auto size (no <code>imageSize</code> / <code>thumbnailSize</code>)</h3>
+<p>
+  Same pattern but without specifying any size inputs. The carousel automatically fills the available modal space. The
+  library adds the necessary flex fixes to the Clarity modal internals without any extra configuration.
+</p>
+
+<button class="btn btn-primary" (click)="openAutoModal()">Open Modal (auto size)</button>
+
+<clr-modal [clrModalOpen]="autoModalOpen()" (clrModalOpenChange)="$event ? null : onAutoModalClose()" clrModalSize="lg">
+  <h3 class="modal-title">
+    <button
+      *ngIf="showAutoCarousel()"
+      class="btn btn-icon btn-link"
+      style="
+        margin: 0 4px 0 0;
+        padding: 0;
+        min-width: 0;
+        width: 24px;
+        height: 24px;
+        display: inline-flex;
+        align-items: center;
+        justify-content: center;
+        vertical-align: middle;
+      "
+      (click)="onAutoCarouselClose()"
+      aria-label="Back"
+    >
+      <cds-icon shape="angle" direction="left" size="24"></cds-icon>
+    </button>
+    <span>Porsche 911 – Images (auto size)</span>
+  </h3>
+
+  <div class="modal-body" style="padding: 0; overflow: hidden; display: flex; flex-direction: column">
+    <!-- Static gallery view – no spotlightSize specified -->
+    <div *ngIf="!showAutoCarousel()" style="padding: 1rem">
+      <clr-image-gallery
+        [images]="images"
+        [productName]="'Porsche 911'"
+        [useBuiltInModal]="false"
+        (galleryOpen)="onAutoGalleryOpen($event)"
+      ></clr-image-gallery>
+    </div>
+
+    <!-- Carousel – no imageSize or thumbnailSize specified -->
+    <clr-image-carousel
+      *ngIf="showAutoCarousel()"
+      [images]="images"
+      [productName]="'Porsche 911'"
+      [initialIndex]="autoCarouselIndex()"
+      [hideHeader]="true"
+      (closeCarousel)="onAutoCarouselClose()"
       style="flex: 1; min-height: 0; display: flex; flex-direction: column"
     ></clr-image-carousel>
   </div>
